refactor: ディレクトリ構成修正
